src/main/java/seedu/address/commons/core/keyword/Keyword.java
package seedu.address.commons.core.keyword;

import seedu.address.model.person.ContainsKeywordsPredicate;

import static java.util.Objects.requireNonNull;

public class Keyword {
    String keyword;

    public Keyword(String keyword) {
        requireNonNull(keyword);
        this.keyword = keyword;
    }

    public boolean isKeywordFound(String string) {
        return string.toLowerCase().contains(keyword.toLowerCase());
    }

    @Override
    public boolean equals(Object other) {
        return other == this // short circuit if same object
                || (other instanceof Keyword // instanceof handles nulls
                && keyword.equals(((Keyword) other).keyword)); // state check
    }

    @Override
    public int hashCode() {
        return keyword.hashCode();
    }
}
